Use DI to inject filters and features
