+++ core/src/main/java/cucumber/runtime/RuntimeOptions.java
@@ -5,14 +5,11 @@ import cucumber.api.SnippetType;
 import cucumber.api.StepDefinitionReporter;
 import cucumber.api.event.EventListener;
 import io.cucumber.datatable.DataTable;
-import cucumber.api.event.TestRunStarted;
 import cucumber.api.formatter.ColorAware;
 import cucumber.api.formatter.Formatter;
 import cucumber.api.formatter.StrictAware;
 import cucumber.runner.EventBus;
 import cucumber.runtime.formatter.PluginFactory;
-import cucumber.runtime.io.ResourceLoader;
-import cucumber.runtime.model.CucumberFeature;
 import cucumber.runtime.model.PathWithLines;
 import cucumber.util.FixJava;
 import cucumber.util.Mapper;
@@ -32,7 +29,6 @@ import java.util.Map;
 import java.util.ResourceBundle;
 import java.util.regex.Pattern;
 
-import static cucumber.runtime.model.CucumberFeature.load;
 import static cucumber.util.FixJava.join;
 import static cucumber.util.FixJava.map;
 import static java.util.Arrays.asList;
